+++ habapp_rules/__init__.py
@@ -2,5 +2,5 @@
 
 import pathlib
 
-__version__ = "7.2.1"
+__version__ = "8.1.2"
 BASE_PATH = pathlib.Path(__file__).parent.parent.resolve()